+++ src/sidecar/sidecarHandle.ts
@@ -153,7 +153,10 @@ export class SidecarHandle {
 
   public getSubjectsV1Api(clusterId: string, connectionId: string): SubjectsV1Api {
     const config: unknown = this.createClientConfig({
-      headers: { "x-cluster-id": clusterId, "x-connection-id": connectionId },
+      headers: {
+        "x-cluster-id": clusterId,
+        "x-connection-id": connectionId,
+      },
     });
     return new SubjectsV1Api(config as SchemaRegistryRestConfiguration);
   }