+++ CHANGELOG.md
@@ -13,6 +13,16 @@ All notable changes to this extension will be documented in this file.
 
 - Topic -> Schema correlation now recognizes schemas using `TopicRecordNameStrategy` in addition to
   `TopicNameStrategy`, [issue #298](https://github.com/confluentinc/vscode/issues/298).
+- The "Local" item in the Resources view now persists regardless of whether or not a local Kafka
+  cluster is discoverable, in preparation for future enhancements to local resource management.
+  [issue #441](https://github.com/confluentinc/vscode/issues/441)
+
+### Fixed
+
+- Temporary disruptions to CCloud auth sessions will now show a notification to the user while the
+  extension sidecar attempts to resolve the issue (instead of invalidating the session immediately
+  and forcing the user to sign in again).
+  [issue #307](https://github.com/confluentinc/vscode/issues/307)
 
 ## 0.18.3