+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this extension will be documented in this file.
 ### Added
 
 - "Open Settings" command to quickly access the Confluent for VS Code extension settings.
+- "Copy URI" context menu item for CCloud Schema Registry items and local Kafka clusters in the
+  Resources view. [issue #445](https://github.com/confluentinc/vscode/issues/445)
 
 ### Changed